Fix `lldap_user` import
Looks like the issue here is that `lldap_user` is validating the
password of a user when it's read. This probably works fine when
actually reading the user from LLDAP. But when attempting to import an
`lldap_user` resource into the Terraform state, it fails.

What's going wrong is that when Terraform attempts to import a resource,
all it has to start with is the `id`. That ought to be enough to import
anything (since the `id` is supposed to uniquely identify the resource
in the actual infrastructure. Unfortunately, because only the `id` is
set, attempting to validate the password means that there's only the
`id` to go off of. When the rest of the logic is evaluated for
validating a password, it fails with a highly cryptic error because
there is no password set:
```
Error: could not bind to ldap server: LDAP Result Code 206 "Empty password not allowed by the client": ldap: empty password not allowed by the client
```

Even if you explicitly set a `password` in the resource you're
attempting to import, you get the same situation (because again,
Terraform doesn't populate those other fields when importing).

We opt to remove the validation from the read-side of things. Validating
still makes sense on the update-side, and the create-side will need the
password in order to create a user. So the other methods should be fine
as they are.

For more information, see
https://github.com/tasansga/terraform-provider-lldap/issues/9#issuecomment-2572121500.