Use `id` as source of truth for reading resources
As it turns out, most of the resources are using data on the resource
itself to read data for the resource. That mostly works for things when
all you're doing is reading resources from LLDAP. But when you need to
import a resource into the Terraform state, it falls apart because
Terraform only provides the `id` as a value (and not the other
attributes).

We move all the resources over to using the `id`, so they can be
importable. This also should continue to work for reading the resources,
since the previously used attributes were either the same as the `id` on
the LLDAP side, or are derived from the `id`.

For more information, see
https://github.com/tasansga/terraform-provider-lldap/issues/9#issuecomment-2572172966.