fix handling for users without password
The provider attempted to set a password even if none was given and failed since LLDAP does not accept empty passwords. With this fix will the provider ignore the user password entirely if it is not set in the resource: Since we cannot check the password in LLDAP, we cannot even check whether it has changed if none is given, so management is not possible.

This also updates the go dependencies and fixes hardcoded paths in the Makefile.

See also: https://github.com/tasansga/terraform-provider-lldap/issues/16 and possibly https://github.com/tasansga/terraform-provider-lldap/issues/13